+++ blockstackers_core/src/randomizer.rs
@@ -1,23 +1,25 @@
+use std::cell::RefCell;
+
 use oorandom::Rand64;
 
 pub struct Randomizer {
-    queue: Vec<i32>,
+    queue: RefCell<Vec<i32>>,
     max: i32,
     current: i32,
-    rng: Rand64
+    rng: RefCell<Rand64>
 }
 
 impl Randomizer {
     pub fn new(max: i32, seed: u128) -> Randomizer {
         let rng = Rand64::new(seed);
-        Randomizer { queue: Vec::new(), max, current: 0, rng }
+        Randomizer { queue: RefCell::new(Vec::new()), max, current: 0, rng: RefCell::new(rng) }
     }
-    pub fn get(&mut self, i: i32) -> i32 {
-        while (self.queue.len() as i32) < i+1 {
-            let new_num = self.rng.rand_range(0..(self.max as u64));
-            self.queue.push(new_num as i32);
+    pub fn get(&self, i: i32) -> i32 {
+        while (self.queue.borrow().len() as i32) < i+1 {
+            let new_num = self.rng.borrow_mut().rand_range(0..(self.max as u64));
+            self.queue.borrow_mut().push(new_num as i32);
         }
-        return *self.queue.get(i as usize).unwrap();
+        return *self.queue.borrow().get(i as usize).unwrap();
     }
     pub fn current_pointer(&self) -> i32 {
         return self.current.clone() - 1
