+++ client/src/enums.rs
@@ -1,0 +1,9 @@
+use blockstackers_core::buyo_game::{BType, BuyoBuyo};
+
+use crate::gamehandler::GameHandler;
+
+pub enum GameState {
+    Gaming(GameHandler<BuyoBuyo, BType>),
+    Menu,
+    LoadingAssets,
+}
